+++ src/main/java/controller/CreateUserController.java
@@ -7,11 +7,11 @@ import org.slf4j.LoggerFactory;
 import webserver.HttpRequest;
 import webserver.HttpResponse;
 
-public class CreateUserController implements Controller{
+public class CreateUserController extends AbstractController{
     private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
 
     @Override
-    public void service(HttpRequest request, HttpResponse response) {
+    public void doPost(HttpRequest request, HttpResponse response) {
         User user = new User(request.getParameter("userId"),
                 request.getParameter("password"),
                 request.getParameter("name"),
