+++ src/main/java/controller/ListUserController.java
@@ -4,17 +4,19 @@ import db.DataBase;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import util.HttpRequestUtils;
 import webserver.HttpRequest;
 import webserver.HttpResponse;
 
 import java.util.Collection;
+import java.util.Map;
 
-public class ListUserController implements Controller {
+public class ListUserController extends AbstractController {
     private static final Logger log = LoggerFactory.getLogger(ListUserController.class);
 
     @Override
-    public void service(HttpRequest request, HttpResponse response) {
-        if (!request.isLogin()) {
+    public void doGet(HttpRequest request, HttpResponse response) {
+        if (!isLogin(request.getHeader("Cookie"))) {
             response.forward("/user/login.html");
             return ;
         }
@@ -32,4 +34,14 @@ public class ListUserController implements Controller {
         sb.append("</table>");
         response.forwardBody(sb.toString());
     }
+
+    private boolean isLogin(String cookieValue) {
+        Map<String, String> cookies = HttpRequestUtils.parseCookies(cookieValue);
+        String logined = cookies.get("logined");
+        if (logined == null) {
+            return false;
+        }
+        return Boolean.parseBoolean(logined);
+    }
+
 }
