+++ src/main/java/controller/LoginController.java
@@ -7,11 +7,11 @@ import org.slf4j.LoggerFactory;
 import webserver.HttpRequest;
 import webserver.HttpResponse;
 
-public class LoginController implements Controller{
+public class LoginController extends AbstractController{
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
     @Override
-    public void service(HttpRequest request, HttpResponse response) {
+    public void doPost(HttpRequest request, HttpResponse response) {
         User user = DataBase.findUserById(request.getParameter("userId"));
         if (user != null && user.login(request.getParameter("password"))) {
             response.addHeader("Set-Cookie", "logined=true");
